[main] Add Integration Tests

+++ visacards-api/src/main/java/com/afonsoqueiros/springbootinduction/visacardsapi/domain/services/VisaCardService.java
@@ -4,6 +4,8 @@ import com.afonsoqueiros.springbootinduction.visacardsapi.domain.VisaCard;
 import com.afonsoqueiros.springbootinduction.visacardsapi.dtos.CreateVisaCard;
 import com.afonsoqueiros.springbootinduction.visacardsapi.dtos.GetVisaCard;
 import com.afonsoqueiros.springbootinduction.visacardsapi.dtos.UpdateVisaCard;
+import com.afonsoqueiros.springbootinduction.visacardsapi.exceptions.ResourceAlreadyAddedException;
+import com.afonsoqueiros.springbootinduction.visacardsapi.exceptions.ResourceNotFoundException;
 import com.afonsoqueiros.springbootinduction.visacardsapi.repository.VisaCardRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,37 +19,50 @@ public class VisaCardService {
     @Autowired
     private VisaCardRepository visaCardRepository;
 
-    public GetVisaCard findById(Long id){
+    public GetVisaCard findById(Long id) throws ResourceNotFoundException {
         Optional<VisaCard> visaCard = this.visaCardRepository.findById(id);
-
-        return VisaCardMapper.INSTANCE.mapVisaCardToGetVisaCard(visaCard);
+        if(visaCard.isEmpty()){
+                throw new ResourceNotFoundException("VisaCard with this ID not exists");
+            }
+        else{
+            return VisaCardMapper.INSTANCE.mapVisaCardToGetVisaCard(visaCard);
+        }
     }
 
-    public GetVisaCard createVisaCard(CreateVisaCard createVisaCard){
-
-        VisaCard visaCard = VisaCardMapper.INSTANCE.mapCreateVisaCardToVisaCard(createVisaCard);
-            if(visaCardRepository.findVisaCardByCardNumber(visaCard.getCardNumber()) == null){
-                long id;
-                id = this.visaCardRepository.saveAndFlush(visaCard).getId();
+    public GetVisaCard createVisaCard(CreateVisaCard createVisaCard) throws ResourceAlreadyAddedException {
+        long id;
+        if(visaCardRepository.findVisaCardByCardNumber(createVisaCard.getCardNumber()) == null){
+            VisaCard visaCard = VisaCardMapper.INSTANCE.mapCreateVisaCardToVisaCard(createVisaCard);
+            id = this.visaCardRepository.saveAndFlush(visaCard).getId();
                 return VisaCardMapper.INSTANCE.mapVisaCardToGetVisaCard(this.visaCardRepository.findById(id));
             }
-        return null;
+            else{
+                throw new ResourceAlreadyAddedException("VisaCard with this number already added ");
+            }
     }
 
-    public ResponseEntity<Object> updateVisaCard(UpdateVisaCard updateVisaCard, Long id){
+    public void updateVisaCard(UpdateVisaCard updateVisaCard, Long id) throws ResourceNotFoundException {
 
         Optional<VisaCard> visaCardToUpdate = this.visaCardRepository.findById(id);
-        if(visaCardToUpdate.isPresent()){
+        if(visaCardToUpdate.isEmpty()){
+            throw new ResourceNotFoundException("VisaCard with this ID not exists");
+        }
+        else{
             VisaCard visaCard = VisaCardMapper.INSTANCE.mapUpdateVisaCardToVisaCard(Optional.ofNullable(updateVisaCard),visaCardToUpdate);
             this.visaCardRepository.save(visaCard);
         }
-        return ResponseEntity.notFound().build();
     }
 
-    public Optional<VisaCard> deleteVisaCard(Long id){
+    public void deleteVisaCard(Long id) throws ResourceNotFoundException {
         Optional<VisaCard> visaCardToDelete = this.visaCardRepository.findById(id);
-        this.visaCardRepository.deleteById(id);
-        return visaCardToDelete;
+
+        if(visaCardToDelete.isEmpty()){
+            throw new ResourceNotFoundException("VisaCard with this ID not exists");
+        }
+        else{
+            this.visaCardRepository.deleteById(id);
+        }
+
     }
 
 }
